Allow Text Wrapping
Added text wrapping within cells so that cells expand vertically when sentences
do not fit horizontally within the cell.

Testing Steps:
  1. Run the application from main.py
  2. Insert the sentence, "The quick brown fox jumps over the lazy dog"
  3. Visually inspect that the cell grows vertically when the sentence in
     inserted into the cell

Type: New Feature
