+++ View/encoding_table.py
@@ -18,15 +18,11 @@ class EncodingTable(QTableWidget):
         self.setColumnCount(4)
 
         self.horizontalHeader().setStretchLastSection(True)
-        self.horizontalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Stretch)
-        self.horizontalHeader().setDefaultSectionSize(60)
-        self.horizontalHeader().setMinimumSectionSize(60)
-        self.horizontalHeader().setMaximumSectionSize(100)
+        self.horizontalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Interactive)
+        self.horizontalHeader().setDefaultSectionSize(100)
         self.verticalHeader().setStretchLastSection(True)
-        self.verticalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Stretch)
-        self.verticalHeader().setDefaultSectionSize(30)
-        self.verticalHeader().setMinimumSectionSize(30)
-        self.setStyleSheet("QTableWidget::item {border: 0px; padding: 10px;}")
+        self.verticalHeader().setSectionResizeMode(QtWidgets.QHeaderView.ResizeToContents)
+        self.verticalHeader().setDefaultSectionSize(70)
 
         # Ensure at least 5 rows are visible at all times.
         self.minimum_visible_rows = 5
@@ -34,9 +30,36 @@ class EncodingTable(QTableWidget):
         self.setMinimumHeight(self.rowHeight(0) * (self.minimum_visible_rows + total_border_height))
 
     def add_column(self):
-        """Increases the column count of the table by 1."""
+        """
+        Increases the column count of the table by 1.
+        """
+
         self.setColumnCount(self.columnCount() + 1)
 
     def add_row(self):
-        """Increases the row count of the table by 1."""
+        """
+        Increases the row count of the table by 1.
+        """
+
         self.setRowCount(self.rowCount() + 1)
+
+    def set_cell_size(self, width, height):
+        """
+        Changes default cell width
+        """
+
+        self.horizontalHeader().setMinimumSectionSize(width)
+        self.verticalHeader().setMinimumSectionSize(height)
+
+    def set_maximum_width(self, width):
+        """
+        Changes default cell height
+        """
+
+        self.horizontalHeader().setMaximumSectionSize(width)
+
+    def set_padding(self, padding):
+        """
+        Changes default padding
+        """
+        self.setStyleSheet("QTableWidget::item { padding: " + padding + "px }")
